Move useDebounce to utils
