+++ src/utils/hooks.ts
@@ -1,0 +1,17 @@
+import { useEffect, useState } from 'react';
+
+// https://dev.to/gabe_ragland/debouncing-with-react-hooks-jci
+export function useDebounce(value: any, delay: number) {
+  const [debouncedValue, setDebouncedValue] = useState(value);
+
+  useEffect(() => {
+    const handler = setTimeout(() => {
+      setDebouncedValue(value);
+    }, delay);
+    return () => {
+      clearTimeout(handler);
+    };
+  }, [value]);
+
+  return debouncedValue;
+};
